requirements.txt
fastapi
uvicorn[standard]
sqlmodel
SQLAlchemy
asyncpg
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pydantic
pydantic-settings
PyJWT